requirements.txt
pillow
argparse
python-resize-image
